web/requirements.txt
alembic==1.0.0
aniso8601==3.0.2
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-JWT-Extended==3.12.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
rope==0.11.0
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
WTForms==2.2.1
psycopg2
gunicorn==19.6.0
qrcode==6.0
Pillow==5.2.0
pyzbar==0.1.7